Alignment code
Alignment Code, working in simulation. Uses 3d solvers plus gyro to align to tag, speed ramping to slow as the robot nears the tag. Incorportated the driveToPose method from vision as well, which uses pathplanner to drive to a point on the field. Also made a button to toggle between left post and right post while aligning. To test, click a in simulation, and it should align to tag 9(keep in mind my offsets are WAY off and you will be thrown to the side, that is just because i tested with a random number)